Typecast `memberful_use_global_snippet` params when updating global marketing content

+++ wordpress/wp-content/plugins/memberful-wp/src/admin.php
@@ -629,7 +629,7 @@ function memberful_wp_global_marketing() {
       update_option( 'memberful_use_global_marketing', true );
       update_option( 'memberful_global_marketing_override', filter_input( INPUT_POST, 'memberful_global_marketing_override', FILTER_SANITIZE_NUMBER_INT ) );
       update_option( 'memberful_global_marketing_content', filter_input( INPUT_POST, 'memberful_global_marketing_content' ) );
-      update_option ( 'memberful_use_global_snippets', isset($_POST['memberful_use_global_snippets']));
+      update_option( 'memberful_use_global_snippets', (int) isset($_POST['memberful_use_global_snippets']));
     } else {
       update_option( 'memberful_use_global_marketing', false );
     }
